backend/requirements.txt
fastapi
uvicorn
python-dotenv
assemblyai
murf
google-genai
python-multipart
pydantic
dotenv